requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
pydantic>=2.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0